% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplot.R
\name{aplot}
\alias{aplot}
\title{Alive points plot}
\usage{
aplot(.dfr)
}
\arguments{
\item{.dfr}{the data frame of the points with alive status and number of neighbours.}
}
\value{
a plot
}
\description{
Alive points plot
}
