% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplot.R
\name{base_plot}
\alias{base_plot}
\title{Base alive plot}
\usage{
base_plot(.dfr)
}
\arguments{
\item{.dfr}{a 2D matrix or a dataframe with x and y dimension.}
}
\value{
a plot
}
\description{
Takes either a 2D matrix or a data frame and plot alive points
}
