% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{mat_dist}
\alias{mat_dist}
\title{Compute l_inf distances from a point to the alive point matrix}
\usage{
mat_dist(p, m)
}
\arguments{
\item{p, }{a point with 2 dimensions}

\item{m, }{a matrix with dimension n * 2}
}
\value{
, a matrix of dimension n * 1
}
\description{
Compute l_inf distances from a point to the alive point matrix
}
