% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_init.R
\name{mat_init}
\alias{mat_init}
\title{Initiate the start matrix}
\usage{
mat_init(.seed = NULL, n = 40, square_sequence = -5:5, ndim = 2)
}
\arguments{
\item{.seed, }{random seed to use}

\item{n, }{number of max points to generate, duplicates are removed.}

\item{square_sequence, }{integer sequence of the form -1:1 to generate the dimensions.}

\item{ndim, }{integer > 0 for the number of dimensions}
}
\value{
a square matrix of integer points
}
\description{
To create the starting alive points (contained in a square).
}
\examples{
mat_init(.seed = 456, n = 40, square_sequence = -5:5, ndim = 2)
}
