% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbours.R
\name{nb_neig}
\alias{nb_neig}
\title{Find the number of alive neighbours}
\usage{
nb_neig(.mat_init, .mat_ordered)
}
\arguments{
\item{.mat_init}{a matrix of the alive points}

\item{.mat_ordered}{a matrix with all the points and neighbours}
}
\value{
an integer : the number of neighbours.
}
\description{
Find the number of alive neighbours
}
