% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_step.R
\name{next_step}
\alias{next_step}
\title{Next step in the game}
\usage{
next_step(.mat_init, nb_ngb = 3)
}
\arguments{
\item{.mat_init, }{a data frame or a matrix with initial points}

\item{nb_ngb, }{the rule for the number of neighbours to have for a cell to be alive}
}
\value{
a data frame
}
\description{
Next step in the game
}
